+++ lib/screens/my_home_page.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:pocket_coach/screens/morning_questions.dart';
 import 'base.dart';
+
 
 class MyHomePage extends StatefulWidget {
   const MyHomePage({Key? key, required this.title}) : super(key: key);
@@ -11,29 +13,45 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  // String selectedPage = '';
 
   @override
   Widget build(BuildContext context) => Scaffold(
     appBar: AppBar(
-      title: Text(widget.title),
+      title: Text(
+        widget.title,
+        style: const TextStyle(
+          fontSize: 25,
+          color: Color.fromARGB(255, 233, 79, 14),
+        ),
+        ),
     ),
     drawer: const NavigationDrawer(),
-    body: Center(
-      child: Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[
-        const Text(
-          // 'Selected page: $selectedPage',
-          'Initial page',
+    body: SafeArea(
+      child: Center(
+        child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: <Widget>[
+      
+          const Padding(
+            padding: EdgeInsets.all(16.0),
+            child: Text(
+              'Пока что в таком варианте: НАДО жать в левом верхнем углу',
+            style: TextStyle(
+            fontSize: 24,
+            color: Color.fromARGB(255, 33, 134, 3),
+          ),
+            ),
+          ),
+      
+          Text(
+            // 'ABC',
+            '',
+            style: Theme.of(context).textTheme.headlineMedium,
+          ),
+        ],
+      ),
         ),
-        Text(
-          '',
-          style: Theme.of(context).textTheme.headlineMedium,
-        ),
-      ],
     ),
-  ),
   );
 }
 
@@ -48,16 +66,31 @@ class NavigationDrawer extends StatelessWidget {
           children: <Widget>[
             const DrawerHeader(
               decoration: BoxDecoration(
-                color: Colors.blue,
+                color: Color.fromARGB(255, 3, 180, 74),
               ),
               child: Text(
-                'Drawer Header',
+                'Меню карманного коуча',
                 style: TextStyle(
                   color: Colors.white,
                   fontSize: 24,
                 ),
               ),
             ),
+
+            ListTile(
+              leading: const Icon(Icons.alarm_on),
+              title: const Text('Утренние вопросы'),
+              onTap: () {
+                Navigator.push(
+                  context, 
+                  MaterialPageRoute(builder: (context){
+                      return const MorningQuestionsScreen();
+                    }
+                  )
+                );
+              },
+            ),
+
             ListTile(
               leading: const Icon(Icons.message),
               title: const Text('Базовые вопросы'),
@@ -72,15 +105,7 @@ class NavigationDrawer extends StatelessWidget {
                 
               },
             ),
-            // ListTile(
-            //   leading: const Icon(Icons.account_circle),
-            //   title: const Text('Profile'),
-            //   onTap: () {
-            //     setState(() {
-            //       selectedPage = 'Profile';
-            //     });
-            //   },
-            // ),
+
             // ListTile(
             //   leading: const Icon(Icons.settings),
             //   title: const Text('Settings'),